requirements.txt
# Azure AI
openai>=1.30.0
azure-search-documents>=11.6.0
azure-ai-contentsafety>=1.0.0
azure-ai-evaluation>=1.0.0
azure-identity>=1.16.0
azure-keyvault-secrets>=4.8.0
azure-storage-blob>=12.20.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Document loading
pypdf>=4.2.0
python-docx>=1.1.0
tiktoken>=0.7.0

# Data handling
pandas>=2.2.0     # CSV ticket loading
numpy>=1.26.0     # cosine similarity for deduplication

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.3.0
httpx>=0.27.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0